:sparkles: [MCH] Add cluster transformation workflow.
Test usage :

o2-mch-clusters-sampler-workflow --nEventsPerTF 1000 -b --infile clusters.O2 |
o2-mch-clusters-transformer-workflow --geometry Detectors/MUON/MCH/Geometry/Test/ideal-geometry-o2.json -b |
o2-mch-clusters-sink-workflow -b --txt --outfile global-clusters.txt --no-digits --global

Also, the clusters- sampler and sink workflow have been squeezed into
one source file per workflow, instead of 3 per workflow previously, to
simplify a bit the Workflow directory that is becoming a bit crowded.
